+++ src/hocon/_main.py
@@ -7,11 +7,22 @@ from .resolver import resolve
 
 
 def load(fp: TextIO) -> ROOT_TYPE:
-    absolute_path = Path.cwd() / fp.name
+    """Load HOCON from an open file.
+
+    Relative path for include and encoding is set automatically.
+    """
+    absolute_path = Path(fp.name).absolute()
     return loads(fp.read(), absolute_path, fp.encoding)
 
 
 def loads(data: str, root_filepath: str | Path | None = None, encoding: str = "UTF-8") -> ROOT_TYPE:
-    root_filepath = root_filepath or Path.cwd()
+    """Load a string to HOCON.
+
+    :param data: string to parse and resolve.
+    :param root_filepath: path to resolve 'include' from. Set current working directory by default.
+    :param encoding: encoding to use for 'include' files.
+    :return: resolved dict or list
+    """
+    root_filepath = root_filepath or Path.cwd() / "application.conf"
     parsed = parse(data, root_filepath=root_filepath, encoding=encoding)
     return resolve(parsed)