+++ tests/acceptance/includes/test_includes.py
@@ -1,9 +1,9 @@
 from pathlib import Path
 
-from hocon import load
+from hocon import loads
 
 
-def test_1():
+def test_2_includes_relative_to_cwd():
     conf_filepath = Path(__file__).parent / "data" / "main.conf"
-    result = load(open(conf_filepath, encoding="UTF-8"))
+    result = loads(conf_filepath.read_text(), encoding="UTF-8")
     assert result == {"house": {"location": "서울 Nice street 42", "city": "서울"}}